ORGANIZEI E ARRUMEI OS BUG  GRACAS A DEUS
