+++ quinzena7/labefy-project/src/components/CreatePlaylist.js
@@ -23,6 +23,7 @@ export default class CreatePlaylistForm extends React.Component {
       }).then((res) => {
         alert("Playlist successfully created", res)
           this.setState({ inputNameValue: "" })
+          this.props.getAllPlaylists()
       }).catch((err) => {
           alert("Error", err.data)
       })
@@ -34,8 +35,6 @@ export default class CreatePlaylistForm extends React.Component {
         <h1>Create your Playlist here!</h1>
         <input value={this.state.inputNameValue} onChange={this.onChangeInputName} placeholder="Playlist name..."></input>
         <button onClick={this.createPlaylist}>+</button>
-        <button onClick={() => this.props.PageChange("PlaylistLibrary")}>Your Library</button>
-        <button onClick={() => this.props.PageChange("Dashboard")}>Back to Dashboard</button>
       </div>
     );
   }
